app/requirements.txt
# requirements.txt
psycopg2-binary
flask